Improve social card images

+++ social_cards_automation_scripts/createImagesFromPosts.sh
@@ -23,44 +23,51 @@ mkdir -p "$IMAGES_DIR"
 # Function to dynamically adjust text to prevent overflow
 adjust_text() {
     local text="$1"
-    local max_length=370  # Max pixel width for text
-    local font_size=44    # Initial font size
+    local max_length="$2"  # Max pixel width for text
+    local font_dir="$3"    # Directory for fonts
+    local font_file="$4"   # Font file name
+    local min_font_size="${5:-20}"  # Minimum font size, default is 20 if not provided
+
+    local font_size=76    # Initial font size
     local adjusted_text="$text"
     local temp_file="temp.png"
+    local text_width
 
     # Use ImageMagick to calculate the width of the rendered text
     while true; do
-        convert -background none -fill white -font "$FONTS_DIR/Work_Sans/WorkSans-VariableFont_wght.ttf" -pointsize $font_size label:"$adjusted_text" "$temp_file"
-        local text_width=$(identify -format "%w" "$temp_file")
+        if ! convert -background none -fill white -font "$font_dir/$font_file" -pointsize $font_size label:"$adjusted_text" "$temp_file"; then
+            echo "Error generating image with convert command."
+            return 1
+        fi
+
+        text_width=$(identify -format "%w" "$temp_file")
         rm -f "$temp_file"
 
-        if [ $text_width -le $max_length ] || [ $font_size -le 20 ]; then
+        if [[ $text_width -le $max_length ]] || [[ $font_size -le $min_font_size ]]; then
             break
         fi
 
-        if [ $text_width -gt $max_length ]; then
-            font_size=$((font_size - 8))  # Reduce font size to make text fit
-        fi
+        font_size=$((font_size - 8))  # Reduce font size to make text fit
     done
 
     # Apply more aggressive wrapping if still necessary
-    if [ $text_width -gt $max_length ]; then
-        adjusted_text=$(echo "$adjusted_text" | fold -w 38 -s)  # Wrap text to fit
+    if [[ $text_width -gt $max_length ]]; then
+        local wrap_width=$((max_length / font_size * 2))  # Dynamic wrap width based on font size
+        adjusted_text=$(echo "$adjusted_text" | fold -w "$wrap_width" -s)  # Wrap text to fit
     fi
 
     echo "$adjusted_text"
 }
 
+
 # Loop through each markdown post file
 for post in "$POSTS_DIR"/*.md; do
     # Extract the title, excerpt, and author
     title=$(awk '/^title:/{gsub(/^title: /,""); gsub(/'"'"'/,""); print; exit}' "$post")
-    excerpt=$(awk '/^excerpt:/{gsub(/^excerpt: /,""); gsub(/'"'"'/,""); print; exit}' "$post")
     author=$(awk '/^author:/{gsub(/^author: /,""); gsub(/'"'"'/,""); print; exit}' "$post")
 
     # Adjust title and excerpt to prevent overflow
-    adjusted_title=$(adjust_text "$title")
-    adjusted_excerpt=$(adjust_text "$excerpt")
+    adjusted_title=$(adjust_text "$title" 275 "$FONTS_DIR" "/Work_Sans/WorkSans-VariableFont_wght.ttf")
 
     # Create an image with the title, excerpt, and branding
     jpg_file="$IMAGES_DIR/$(basename "$post" .md).jpg"
@@ -68,16 +75,14 @@ for post in "$POSTS_DIR"/*.md; do
 
     # Create an image with the title, excerpt, and branding
     convert -size 1200x627 xc:"#9900FC" \
-            \( -size 900x400 -background none -fill white -font "$FONTS_DIR/Work_Sans/WorkSans-VariableFont_wght.ttf" -pointsize 44 \
-               label:"$adjusted_title" -gravity north -geometry +0+100 \) -composite \
-            \( -size 900x600 -background none -fill white -font "$FONTS_DIR/Work_Sans/WorkSans-VariableFont_wght.ttf" -pointsize 36 \
-               label:"$adjusted_excerpt" -gravity center -geometry +0+280 \) -composite \
+            \( -size 900x500 -background none -fill white -font "$FONTS_DIR/Work_Sans/WorkSans-VariableFont_wght.ttf" -pointsize 64 \
+               label:"$adjusted_title" -gravity center \) -geometry +0+60  -composite \
             -font "$FONTS_DIR/Work_Sans/WorkSans-VariableFont_wght.ttf" -pointsize 34 -fill white \
             -gravity southeast -annotate +30+30 "$author" \
             "$IMAGES_DIR/$(basename "$post" .md).jpg"
 
     # Convert JPG to WebP
-    cwebp -q 80 "$jpg_file" -o "$webp_file"
+    cwebp -q 100 "$jpg_file" -o "$webp_file"
 
 done
